db/7_semester_project/datingapp/requirements.txt
Django==3.2.6
psycopg2-binary
gunicorn==20.1.0
django-cors-headers==3.5.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.1
PyJWT==1.7.1
